[BO] #3 스키마 변경에 따른 문서 자료 업데이트
- 파일 관련 테이블 스키마 중 size 관련 컬럼 타입 변경 (integer -> bigint)
- README 내 erd 업데이트

+++ admin-api/db/schema.sql
@@ -45,7 +45,7 @@ create table thumbnail_image
 (
     id          bigint primary key   default nextval('seq_thumbnail_image'),
     file_url    text unique not null,
-    file_size   integer     not null default 0,
+    file_size   bigint      not null default 0,
     created_at  timestamp   not null default current_timestamp,
     modified_at timestamp   not null default current_timestamp
 );
